+++ api_gateway/app/api.py
@@ -4,7 +4,6 @@ from contextlib import asynccontextmanager
 import climatoology
 import uvicorn
 import yaml
-from climatoology.app.platform import CeleryPlatform
 from fastapi import FastAPI
 from fastapi_cache import FastAPICache
 from fastapi_cache.backends.inmemory import InMemoryBackend
@@ -12,6 +11,7 @@ from fastapi_cache.backends.inmemory import InMemoryBackend
 import api_gateway
 from api_gateway.app.route import computation, health, metadata, plugin, store
 from api_gateway.app.settings import GatewaySettings
+from api_gateway.sender import CelerySender
 
 settings = GatewaySettings()
 
@@ -51,7 +51,7 @@ It servers as the single point of interaction.
 async def configure_dependencies(the_app: FastAPI):
     log.debug('Configuring Platform connection')
     the_app.state.settings = settings
-    the_app.state.platform = CeleryPlatform()
+    the_app.state.platform = CelerySender()
     FastAPICache.init(InMemoryBackend())
     log.debug('Platform configured')
     yield